+++ SwiftyButton/SwiftyButton+ResizableImage.swift
@@ -10,26 +10,26 @@ import UIKit
 
 extension SwiftyButton {
     
-    class func buttonImageWithColor(
+    class func buttonImage(
         color: UIColor,
         shadowHeight: CGFloat,
         shadowColor: UIColor,
         cornerRadius: CGFloat) -> UIImage {
         
-        return buttonImageWithColor(color, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, frontImageOffset: 0)
+        return buttonImage(color: color, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, frontImageOffset: 0)
     }
     
-    class func highlightedButtonImageWithColor(
+    class func highlightedButtonImage(
         color: UIColor,
         shadowHeight: CGFloat,
         shadowColor: UIColor,
         cornerRadius: CGFloat,
         buttonPressDepth: CGFloat) -> UIImage {
         
-        return buttonImageWithColor(color, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, frontImageOffset: shadowHeight * buttonPressDepth)
+        return buttonImage(color: color, shadowHeight: shadowHeight, shadowColor: shadowColor, cornerRadius: cornerRadius, frontImageOffset: shadowHeight * buttonPressDepth)
     }
     
-    private class func buttonImageWithColor(
+    private class func buttonImage(
         color: UIColor,
         shadowHeight: CGFloat,
         shadowColor: UIColor,
@@ -41,37 +41,37 @@ extension SwiftyButton {
         let height     = max(1, cornerRadius * 2 + shadowHeight)
         let size       = CGSize(width: width, height: height)
             
-        let frontImage = imageWithColor(color, size: size, cornerRadius: cornerRadius)
+        let frontImage = image(color: color, size: size, cornerRadius: cornerRadius)
         var backImage: UIImage? = nil
         if shadowHeight != 0 {
-            backImage = imageWithColor(shadowColor, size: size, cornerRadius: cornerRadius)
+            backImage = image(color: shadowColor, size: size, cornerRadius: cornerRadius)
         }
         
         let rect = CGRect(x: 0, y: 0, width: size.width, height: size.height + shadowHeight)
         
         // Draw background image then foreground image
         UIGraphicsBeginImageContextWithOptions(rect.size, false, 0)
-        backImage?.drawAtPoint(CGPoint(x: 0, y: shadowHeight))
-        frontImage.drawAtPoint(CGPoint(x: 0, y: frontImageOffset))
+        backImage?.draw(at: CGPoint(x: 0, y: shadowHeight))
+        frontImage.draw(at: CGPoint(x: 0, y: frontImageOffset))
         let nonResizableImage = UIGraphicsGetImageFromCurrentImageContext()
         UIGraphicsEndImageContext()
         
         // Create resizable image
         let capInsets = UIEdgeInsets(top: cornerRadius + frontImageOffset, left: cornerRadius, bottom: cornerRadius + shadowHeight - frontImageOffset, right: cornerRadius)
-        let resizableImage = nonResizableImage.resizableImageWithCapInsets(capInsets, resizingMode: .Stretch)
+        let resizableImage = nonResizableImage?.resizableImage(withCapInsets: capInsets, resizingMode: .stretch)
             
-        return resizableImage
+        return resizableImage!
     }
     
-    private class func imageWithColor(color: UIColor, size: CGSize, cornerRadius: CGFloat) -> UIImage {
+    private class func image(color: UIColor, size: CGSize, cornerRadius: CGFloat) -> UIImage {
         
-        let rect = CGRect(origin: CGPointMake(0, 0), size: size)
+        let rect = CGRect(origin: CGPoint(x: 0, y: 0), size: size)
         
         // Create a non-rounded image
     	UIGraphicsBeginImageContextWithOptions(size, false, 0)
         let context = UIGraphicsGetCurrentContext()
-        CGContextSetFillColorWithColor(context, color.CGColor)
-        CGContextFillRect(context, rect)
+        context?.setFillColor(color.cgColor)
+        context?.fill(rect)
         let nonRoundedImage = UIGraphicsGetImageFromCurrentImageContext()
         UIGraphicsEndImageContext()
         
@@ -81,11 +81,11 @@ extension SwiftyButton {
             roundedRect: rect,
             cornerRadius: cornerRadius
         ).addClip()
-        nonRoundedImage.drawInRect(rect)
+        nonRoundedImage?.draw(in: rect)
         
         let image = UIGraphicsGetImageFromCurrentImageContext()
         UIGraphicsEndImageContext()
         
-        return image
+        return image!
     }
 }